# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Castro Digital Platform API","description":"Current backend foundation contract; business rules not represented here are not implied.","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/requests":{"post":{"tags":["request-controller"],"operationId":"create","parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestInput"}}},"required":true},"responses":{"200":{"description":"Request accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RequestResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}}}},"/api/v1/bookings":{"post":{"tags":["booking-controller"],"operationId":"create_1","parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingRequest"}}},"required":true},"responses":{"200":{"description":"Booking created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"200":{"description":"Session ended","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthLogoutResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginInput"}}},"required":true},"responses":{"200":{"description":"Session established","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthLoginResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}}}},"/api/v1/spaces":{"get":{"tags":["public-catalog-controller"],"operationId":"spaces","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SpaceItem"}}}}}}}},"/api/v1/spaces/{slug}":{"get":{"tags":["public-catalog-controller"],"operationId":"space","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SpaceItem"}}}}}}},"/api/v1/services":{"get":{"tags":["public-catalog-controller"],"summary":"List active services","operationId":"services","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CatalogItem"}}}}}}}},"/api/v1/services/{slug}":{"get":{"tags":["public-catalog-controller"],"operationId":"service","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CatalogItem"}}}}}}},"/api/v1/public/config":{"get":{"tags":["public-config-controller"],"operationId":"config","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PublicConfigResponse"}}}}}}},"/api/v1/courses":{"get":{"tags":["public-catalog-controller"],"operationId":"courses","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseItem"}}}}}}}},"/api/v1/courses/{slug}":{"get":{"tags":["public-catalog-controller"],"operationId":"course","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourseItem"}}}}}}},"/api/v1/courses/{id}/sessions":{"get":{"tags":["public-catalog-controller"],"operationId":"sessions","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseSessionResponse"}}}}}}}},"/api/v1/bookings/{reference}":{"get":{"tags":["booking-controller"],"operationId":"lookup","parameters":[{"name":"reference","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Booking found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PublicBookingLookup"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}}}},"/api/v1/availability":{"get":{"tags":["availability-controller"],"operationId":"slots","parameters":[{"name":"bookableType","in":"query","required":true,"schema":{"type":"string","enum":["SERVICE","SPACE","COURSE_SESSION","CONSULTATION"]}},{"name":"bookableId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"durationMinutes","in":"query","required":true,"schema":{"type":"integer","format":"int32","minimum":1}}],"responses":{"200":{"description":"Availability calculated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AvailabilityResult"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"Current session","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthMeResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"security":[{"sessionCookie":[]}]}},"/api/v1/auth/csrf":{"get":{"tags":["auth-controller"],"operationId":"csrf","parameters":[{"name":"token","in":"query","required":true,"schema":{"$ref":"#/components/schemas/CsrfToken"}}],"responses":{"200":{"description":"CSRF token delivered in XSRF-TOKEN cookie","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CsrfTokenResponse"}}}}}}}},"components":{"schemas":{"RequestInput":{"type":"object","properties":{"firstName":{"type":"string","minLength":1},"lastName":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"phone":{"type":"string"},"type":{"type":"string","enum":["CONSULTATION","CORPORATE_PROPOSAL","TRAINING_INFO","SPACE_INFO","GENERAL"]},"message":{"type":"string"}},"required":["email","firstName","lastName","type"]},"ProblemDetailResponse":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"},"timestamp":{"type":"string","format":"date-time"},"details":{"type":"object","additionalProperties":{}}}},"RequestResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"status":{"type":"string"}}},"BookingRequest":{"type":"object","properties":{"bookableType":{"type":"string","enum":["SERVICE","SPACE","COURSE_SESSION","CONSULTATION"]},"bookableId":{"type":"string","format":"uuid"},"date":{"type":"string","format":"date"},"startTime":{"type":"string"},"endTime":{"type":"string"},"participants":{"type":"integer","format":"int32"},"customer":{"$ref":"#/components/schemas/CustomerInput"},"spaceConfiguration":{"$ref":"#/components/schemas/SpaceConfiguration"},"notes":{"type":"string"}},"required":["bookableId","bookableType","customer","date","endTime","startTime"]},"CustomerInput":{"type":"object","properties":{"firstName":{"type":"string","minLength":1},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"}},"required":["firstName"]},"SpaceConfiguration":{"type":"object","properties":{"layoutId":{"type":"string","format":"uuid"},"purpose":{"type":"string"},"amenityIds":{"type":"array","items":{"type":"string","format":"uuid"}}}},"BookingResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"reference":{"type":"string"},"status":{"type":"string"},"startAt":{"type":"string","format":"date-time"},"endAt":{"type":"string","format":"date-time"}}},"AuthLogoutResponse":{"type":"object","properties":{"loggedOut":{"type":"boolean"}}},"LoginInput":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"AuthLoginResponse":{"type":"object","properties":{"email":{"type":"string"},"authenticated":{"type":"boolean"}}},"SpaceItem":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"slug":{"type":"string"},"description":{"type":"string"},"location":{"type":"string"},"capacityMin":{"type":"integer","format":"int32"},"capacityMax":{"type":"integer","format":"int32"}}},"CatalogItem":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"slug":{"type":"string"},"description":{"type":"string"},"durationMinutes":{"type":"integer","format":"int32"},"bookingEnabled":{"type":"boolean"}}},"PublicConfigResponse":{"type":"object","properties":{"businessTimezone":{"type":"string"}}},"CourseItem":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"slug":{"type":"string"},"description":{"type":"string"}}},"CourseSessionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"startAt":{"type":"string","format":"date-time"},"endAt":{"type":"string","format":"date-time"}}},"PublicBookingLookup":{"type":"object","properties":{"reference":{"type":"string"},"status":{"type":"string"},"startAt":{"type":"string","format":"date-time"},"endAt":{"type":"string","format":"date-time"}}},"AvailabilityResult":{"type":"object","properties":{"date":{"type":"string","format":"date"},"timezone":{"type":"string"},"slots":{"type":"array","items":{"$ref":"#/components/schemas/AvailabilitySlot"}}}},"AvailabilitySlot":{"type":"object","properties":{"start":{"type":"string"},"end":{"type":"string"},"status":{"type":"string"}}},"AuthMeResponse":{"type":"object","properties":{"email":{"type":"string"},"authenticated":{"type":"boolean"}}},"CsrfToken":{"type":"object","properties":{"parameterName":{"type":"string"},"token":{"type":"string"},"headerName":{"type":"string"}}},"CsrfTokenResponse":{"type":"object","properties":{"token":{"type":"string"},"headerName":{"type":"string"},"parameterName":{"type":"string"}}}},"securitySchemes":{"sessionCookie":{"type":"apiKey","description":"Session cookie established by login. State-changing browser requests also require the XSRF-TOKEN cookie and X-XSRF-TOKEN header.","name":"JSESSIONID","in":"cookie"}}}}